+++ plugins/flutter_aepmessaging/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 4.0.1
+
+* Add namespace support for Android
+* Update Flutter support version to 3.x
+
+## 4.0.0
+
+* Adobe Mobile SDK for iOS 5.x support
+* Adobe Mobile SDK for Android 3.x support
+
 ## 3.0.0-beta.1
 
 * Initial release candidate